fix(models): ensure ROW_NUMBER returns int4 type in capacity models

+++ models/transformations/thematiques/capacite/cap_cb_pbo.sql
@@ -29,7 +29,7 @@ WITH destination AS (
 )
 
 SELECT
-    ROW_NUMBER() OVER (ORDER BY cb.cb_code, ds.bp_code) AS id,
+    ROW_NUMBER() OVER (ORDER BY cb.cb_code, ds.bp_code)::int4 AS id,
     cb.cb_code,
     cb.cb_capafo,
     ds.bp_code,
